Add some automated tests for seccomp-sandbox

This covers some of the syscalls that are proxied by the trusted process.

This adds a basic test runner framework.

Fix error return code for open() and other syscalls on x86-64

On error, the open() syscall was returning errno & 0xffffffff in %rax
on x86-64 instead of -errno.  This stops glibc from regarding this as
an error, and so its syscall wrapper returns -errno instead of -1 and
does not set errno.

I have fixed up the other syscalls to use long (64-bit) instead of int
(32-bit) as well.  Not all of them were affected by the problem: it
depends on gcc's code generation.  Sometimes casting to int truncates
a value, sometimes it doesn't.  It seems better to be consistent
though.

Adds a test for open() and some other syscalls.

TODO: Need to figure out how to run the tests automatically

BUG=none
TEST=none
Review URL: http://codereview.chromium.org/1729003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@45379 0039d316-1c4b-4281-b951-d872f2087c98

+++ sandbox/linux/seccomp/access.cc
@@ -8,7 +8,7 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 
 namespace playground {
 
-int Sandbox::sandbox_access(const char *pathname, int mode) {
+long Sandbox::sandbox_access(const char *pathname, int mode) {
   long long tm;
   Debug::syscall(&tm, __NR_access, "Executing handler");
   size_t len                      = strlen(pathname);
@@ -33,7 +33,7 @@ int Sandbox::sandbox_access(const char *pathname, int mode) {
     die("Failed to forward access() request [sandbox]");
   }
   Debug::elapsed(tm, __NR_access);
-  return static_cast<int>(rc);
+  return rc;
 }
 
 bool Sandbox::process_access(int parentMapsFd, int sandboxFd, int threadFdPub,
