+++ sandbox/linux/seccomp/clone.cc
@@ -8,8 +8,8 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 
 namespace playground {
 
-int Sandbox::sandbox_clone(int flags, char* stack, int* pid, int* ctid,
-                           void* tls, void *wrapper_sp) {
+long Sandbox::sandbox_clone(int flags, char* stack, int* pid, int* ctid,
+                            void* tls, void *wrapper_sp) {
   long long tm;
   Debug::syscall(&tm, __NR_clone, "Executing handler");
   struct {
@@ -98,7 +98,7 @@ int Sandbox::sandbox_clone(int flags, char* stack, int* pid, int* ctid,
     }
   }
   Debug::elapsed(tm, __NR_clone);
-  return static_cast<int>(rc);
+  return rc;
 }
 
 bool Sandbox::process_clone(int parentMapsFd, int sandboxFd, int threadFdPub,
