+++ sandbox/linux/seccomp/exit.cc
@@ -8,7 +8,7 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 
 namespace playground {
 
-int Sandbox::sandbox_exit(int status) {
+long Sandbox::sandbox_exit(int status) {
   long long tm;
   Debug::syscall(&tm, __NR_exit, "Executing handler");
   struct {
