+++ sandbox/linux/seccomp/getpid.cc
@@ -8,7 +8,7 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 
 namespace playground {
 
-int Sandbox::sandbox_getpid() {
+long Sandbox::sandbox_getpid() {
   long long tm;
   Debug::syscall(&tm, __NR_getpid, "Executing handler");
   Debug::elapsed(tm, __NR_getpid);
