+++ sandbox/linux/seccomp/ioctl.cc
@@ -8,7 +8,7 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 
 namespace playground {
 
-int Sandbox::sandbox_ioctl(int d, int req, void *arg) {
+long Sandbox::sandbox_ioctl(int d, int req, void *arg) {
   long long tm;
   Debug::syscall(&tm, __NR_ioctl, "Executing handler");
   struct {
@@ -30,7 +30,7 @@ int Sandbox::sandbox_ioctl(int d, int req, void *arg) {
     die("Failed to forward ioctl() request [sandbox]");
   }
   Debug::elapsed(tm, __NR_ioctl);
-  return static_cast<int>(rc);
+  return rc;
 }
 
 bool Sandbox::process_ioctl(int parentMapsFd, int sandboxFd, int threadFdPub,
