+++ sandbox/linux/seccomp/madvise.cc
@@ -8,7 +8,7 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 
 namespace playground {
 
-int Sandbox::sandbox_madvise(void* start, size_t length, int advice) {
+long Sandbox::sandbox_madvise(void* start, size_t length, int advice) {
   long long tm;
   Debug::syscall(&tm, __NR_madvise, "Executing handler");
   struct {
@@ -30,7 +30,7 @@ int Sandbox::sandbox_madvise(void* start, size_t length, int advice) {
     die("Failed to forward madvise() request [sandbox]");
   }
   Debug::elapsed(tm, __NR_madvise);
-  return static_cast<int>(rc);
+  return rc;
 }
 
 bool Sandbox::process_madvise(int parentMapsFd, int sandboxFd, int threadFdPub,
