+++ sandbox/linux/seccomp/mprotect.cc
@@ -8,7 +8,7 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 
 namespace playground {
 
-int Sandbox::sandbox_mprotect(const void *addr, size_t len, int prot) {
+long Sandbox::sandbox_mprotect(const void *addr, size_t len, int prot) {
   long long tm;
   Debug::syscall(&tm, __NR_mprotect, "Executing handler");
   struct {
@@ -30,7 +30,7 @@ int Sandbox::sandbox_mprotect(const void *addr, size_t len, int prot) {
     die("Failed to forward mprotect() request [sandbox]");
   }
   Debug::elapsed(tm, __NR_mprotect);
-  return static_cast<int>(rc);
+  return rc;
 }
 
 bool Sandbox::process_mprotect(int parentMapsFd, int sandboxFd,
