+++ sandbox/linux/seccomp/munmap.cc
@@ -8,7 +8,7 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 
 namespace playground {
 
-int Sandbox::sandbox_munmap(void* start, size_t length) {
+long Sandbox::sandbox_munmap(void* start, size_t length) {
   long long tm;
   Debug::syscall(&tm, __NR_munmap, "Executing handler");
   struct {
@@ -29,7 +29,7 @@ int Sandbox::sandbox_munmap(void* start, size_t length) {
     die("Failed to forward munmap() request [sandbox]");
   }
   Debug::elapsed(tm, __NR_munmap);
-  return static_cast<int>(rc);
+  return rc;
 }
 
 bool Sandbox::process_munmap(int parentMapsFd, int sandboxFd, int threadFdPub,
