+++ sandbox/linux/seccomp/open.cc
@@ -8,7 +8,7 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 
 namespace playground {
 
-int Sandbox::sandbox_open(const char *pathname, int flags, mode_t mode) {
+long Sandbox::sandbox_open(const char *pathname, int flags, mode_t mode) {
   long long tm;
   Debug::syscall(&tm, __NR_open, "Executing handler");
   size_t len                    = strlen(pathname);
@@ -34,7 +34,7 @@ int Sandbox::sandbox_open(const char *pathname, int flags, mode_t mode) {
     die("Failed to forward open() request [sandbox]");
   }
   Debug::elapsed(tm, __NR_open);
-  return static_cast<int>(rc);
+  return rc;
 }
 
 bool Sandbox::process_open(int parentMapsFd, int sandboxFd, int threadFdPub,
