+++ sandbox/linux/seccomp/stat.cc
@@ -8,7 +8,7 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 
 namespace playground {
 
-int Sandbox::sandbox_stat(const char *path, void *buf) {
+long Sandbox::sandbox_stat(const char *path, void *buf) {
   long long tm;
   Debug::syscall(&tm, __NR_stat, "Executing handler");
   size_t len                    = strlen(path);
@@ -34,10 +34,10 @@ int Sandbox::sandbox_stat(const char *path, void *buf) {
     die("Failed to forward stat() request [sandbox]");
   }
   Debug::elapsed(tm, __NR_stat);
-  return static_cast<int>(rc);
+  return rc;
 }
 
-int Sandbox::sandbox_lstat(const char *path, void *buf) {
+long Sandbox::sandbox_lstat(const char *path, void *buf) {
   long long tm;
   Debug::syscall(&tm, __NR_lstat, "Executing handler");
   size_t len                    = strlen(path);
@@ -63,11 +63,11 @@ int Sandbox::sandbox_lstat(const char *path, void *buf) {
     die("Failed to forward lstat() request [sandbox]");
   }
   Debug::elapsed(tm, __NR_lstat);
-  return static_cast<int>(rc);
+  return rc;
 }
 
 #if defined(__NR_stat64)
-int Sandbox::sandbox_stat64(const char *path, void *buf) {
+long Sandbox::sandbox_stat64(const char *path, void *buf) {
   long long tm;
   Debug::syscall(&tm, __NR_stat64, "Executing handler");
   size_t len                    = strlen(path);
@@ -93,10 +93,10 @@ int Sandbox::sandbox_stat64(const char *path, void *buf) {
     die("Failed to forward stat64() request [sandbox]");
   }
   Debug::elapsed(tm, __NR_stat64);
-  return static_cast<int>(rc);
+  return rc;
 }
 
-int Sandbox::sandbox_lstat64(const char *path, void *buf) {
+long Sandbox::sandbox_lstat64(const char *path, void *buf) {
   long long tm;
   Debug::syscall(&tm, __NR_lstat64, "Executing handler");
   size_t len                    = strlen(path);
@@ -122,7 +122,7 @@ int Sandbox::sandbox_lstat64(const char *path, void *buf) {
     die("Failed to forward lstat64() request [sandbox]");
   }
   Debug::elapsed(tm, __NR_lstat64);
-  return static_cast<int>(rc);
+  return rc;
 }
 #endif
